+++ config/initializers/simple_form_bootstrap.rb
@@ -146,16 +146,6 @@ SimpleForm.setup do |config|
     b.use :hint,  wrap_with: { tag: 'small', class: 'form-text' }
   end
 
-  config.wrappers :multi_select, tag: 'div', class: 'form-group', error_class: 'text-danger has-danger' do |b|
-    b.use :html5
-    b.optional :readonly
-    b.use :label, class: 'control-label'
-    b.wrapper tag: 'div', class: 'form-inline' do |ba|
-      ba.use :input, class: 'form-control'
-      ba.use :error, wrap_with: { tag: 'small', class: 'form-text' }
-      ba.use :hint,  wrap_with: { tag: 'small', class: 'form-text' }
-    end
-  end
   # Wrappers for forms and inputs using the Bootstrap toolkit.
   # Check the Bootstrap docs (http://getbootstrap.com)
   # to learn about the different styles for forms and inputs,
@@ -166,8 +156,5 @@ SimpleForm.setup do |config|
     radio_buttons: :vertical_radio_and_checkboxes,
     file: :vertical_file_input,
     boolean: :vertical_boolean,
-    datetime: :multi_select,
-    date: :multi_select,
-    time: :multi_select
   }
 end