+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 2019-01-29
+### Changed
+- Changed datetime fields to use date picker by default, removed `date_picker` input.
+
 ## 2019-01-28
 ### Added
 - Added `folio_console_dashboard_redirect` config